# backend/docs/swagger.yaml
basePath: /
definitions:
  main.Book:
    properties:
      author:
        type: string
      id:
        type: integer
      title:
        type: string
      year:
        type: integer
    type: object
  main.errorResponse:
    properties:
      error:
        type: string
    type: object
  main.processURLRequest:
    properties:
      operation:
        type: string
      url:
        type: string
    type: object
  main.processURLResponse:
    properties:
      processed_url:
        type: string
    type: object
info:
  contact: {}
  description: Books CRUD + URL Processor service
  title: byFood Assignment API
  version: "1.0"
paths:
  /books:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Book'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: List all books
      tags:
      - books
    post:
      consumes:
      - application/json
      parameters:
      - description: Book
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/main.Book'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Create a new book
      tags:
      - books
  /books/{id}:
    delete:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Delete a book by ID
      tags:
      - books
    get:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Get a book by ID
      tags:
      - books
    put:
      consumes:
      - application/json
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Book
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/main.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Update a book by ID
      tags:
      - books
  /process-url:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.processURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.processURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Process a URL (canonical/redirection/all)
      tags:
      - url
swagger: "2.0"